fix type error
